Encrypted HLS playback fix
when the VGDRM flag was enabled by defualt on phoenix HLS encrypted content playback would crash CGMI. This was because the switch
between the different DRM types in the HLS plugin were done with compilation flags instead of drm Type information on the URI passed down the gst chain. The compilation flags were removed, the drm Type was parsed and cached in the HLS plugin and used when setitng
the caps.